Changes number of instances to int.

+++ JAICore/jaicore-ml/src/jaicore/ml/tsc/dataset/TimeSeriesDataset.java
@@ -180,7 +180,7 @@ public class TimeSeriesDataset {
         return valueMatrices.size();
     }
 
-    public long getNumberOfInstances() {
+    public int getNumberOfInstances() {
         return numberOfInstances;
     }
